Minor improvements

+++ src/util.f90
@@ -1,16 +1,16 @@
 module util
     use iso_fortran_env, only: dp => real64
-    
+
     implicit none
 
 contains
 
     subroutine linspace(a, b, n, x)
-        real(8), intent(in) :: a, b
+        real(dp), intent(in) :: a, b
         integer, intent(in) :: n
-        real(8), dimension(:), intent(out) :: x
+        real(dp), dimension(:), intent(out) :: x
 
-        real(8) :: dx
+        real(dp) :: dx
         integer :: i
 
         dx = (b - a) / (n - 1)
